part 1 of making sure commands that are deletable are deleted

+++ src/commands/fun/compliment.ts
@@ -50,7 +50,7 @@ export class Compliment extends HttpClient {
 
     const message = this.createMessage(command, complimentObj.compliment);
 
-    await command.delete();
+    if (command.deletable) await command.delete();
     return message.startsWith('<')
       ? command.channel.send(message)
       : command.reply(message);
